+++ opencascade-hs/cpp/hs_BRepGProp.cpp
@@ -1,0 +1,6 @@
+#include <BRepGProp.hxx>
+#include "hs_BRepGProp.h"
+
+void hs_BRepGProp_VolumeProperties(TopoDS_Shape *shape, GProp_GProps *props, bool onlyClosed, bool skipShared, bool useTriangulation ){
+    BRepGProp::VolumeProperties(*shape, *props, onlyClosed, skipShared, useTriangulation);
+}