+++ opencascade-hs/cpp/hs_types.h
@@ -62,6 +62,7 @@ typedef int Graphic3d_VerticalTextAlignment;
 typedef void BRepOffsetAPI_MakeOffsetShape;
 typedef int BRepOffset_Mode;
 typedef int GeomAbs_JoinType;
+typedef void GProp_GProps;
 #define Handle(X) void
 #define ARRAY_1(X) void
 #else // __cplusplus