+++ waterfall-cad/src/Waterfall/Solids.hs
@@ -9,24 +9,37 @@ module Waterfall.Solids
 , centeredCylinder
 , unitCone
 , prism
+, volume
+, centerOfMass
+, momentOfInertia
 ) where
 
 
-import Waterfall.Internal.Solid (Solid (..), solidFromAcquire, nowhere)
-import Waterfall.Internal.Finalizers (toAcquire)
+import Waterfall.Internal.Solid (Solid (..), solidFromAcquire, acquireSolid, nowhere)
+import Waterfall.Internal.Finalizers (toAcquire, unsafeFromAcquire)
 import Waterfall.TwoD.Internal.Shape (rawShape)
+import Waterfall.Internal.Edges (gpPntToV3)
 import Waterfall.Transforms (translate)
 import qualified Waterfall.TwoD.Shape as TwoD.Shape
 import qualified OpenCascade.BRepPrimAPI.MakeBox as MakeBox
 import qualified OpenCascade.BRepPrimAPI.MakeSphere as MakeSphere
 import qualified OpenCascade.BRepPrimAPI.MakeCylinder as MakeCylinder
 import qualified OpenCascade.BRepPrimAPI.MakeCone as MakeCone
+import qualified OpenCascade.GProp.GProps as GProps
+import qualified OpenCascade.BRepGProp as BRepGProp
 import qualified OpenCascade.GP as GP
-import Linear (V3 (..), unit, _z, (^*))
+import Control.Lens ((^.))
+import Linear (V3 (..), unit, _x, _y, _z, (^*))
 import qualified OpenCascade.GP.Pnt as GP.Pnt
 import qualified OpenCascade.GP.Vec as GP.Vec
+import qualified OpenCascade.GP.Dir as GP.Dir
+import qualified OpenCascade.GP.Ax1 as GP.Ax1
 import qualified OpenCascade.BRepPrimAPI.MakePrism as MakePrism
 import qualified OpenCascade.Inheritance as Inheritance
+import Control.Monad.IO.Class (liftIO)
+import Control.Monad ((<=<))
+import Foreign.Ptr (Ptr)
+import Data.Acquire (Acquire)
 
 -- | A cube with side lengths of 1, one vertex on the origin, another on \( (1, 1, 1) \)
 unitCube :: Solid
@@ -83,5 +96,28 @@ prism len face = solidFromAcquire $ do
     v <- GP.Vec.new 0 0 len
     MakePrism.fromVec p v True True
 
+gPropQuery :: (Ptr GProps.GProps -> Acquire a) -> Solid -> a
+gPropQuery f s = unsafeFromAcquire $ do
+    solid <- acquireSolid s
+    gProp <- GProps.new
+    liftIO $ BRepGProp.volumeProperties solid gProp False False False
+    f gProp
 
+-- | Volume of the Solid
+volume :: Solid -> Double
+volume = gPropQuery (liftIO . GProps.mass)
 
+-- | Center Of Mass of the Solid
+centerOfMass :: Solid -> V3 Double 
+centerOfMass = gPropQuery ((liftIO . gpPntToV3) <=< GProps.centreOfMass)
+
+-- | Moment of Inertia of the Solid 
+momentOfInertia :: V3 Double -- ^ Point on the Axis of the Moment
+    -> V3 Double -- ^ Direction of the Axis of the Moment 
+    -> Solid
+    -> Double 
+momentOfInertia center axis = gPropQuery $ \gprop -> do
+    pnt <- GP.Pnt.new (center ^. _x) (center ^. _y) (center ^. _z)
+    dir <- GP.Dir.new (axis ^. _x) (axis ^. _y) (axis ^. _z)
+    ax1 <- GP.Ax1.new pnt dir
+    liftIO $ GProps.momentOfInertia gprop ax1